Major update - see description
Fixed a crucial error with the writing of stat files, that would cause the filesize to increase exponentially and thus freeze everything up.
Added the functionality to swipe again at the markDone screen along with a 10 second threaded timer that will return to the main screen if no further action is taken.
Added a sendMail type to notify the cashier everytime a user is created with manually input balance.
Added argument descriptions to the backup function.