Moved data locations and made a new tool
Moved all "permanent" data files to a new folder ./Data/Permanent, and made changes accordingly.
The refUsers.csv file now belong in the ./Data/Users folder.
Added a payment mode to the backup script.
Added the autoPayment tool. It is not commented, but it makes a very thorough log/terminal output, so it shouldn't be hard to follow.