Dialog boxes, cleaning and fixing
Found out that dialog boxes center automatically as long as they are not manually moved (so all the "move"s were removed).
Cleaned the scripts a bit, added a messageBox superclass that changes the background color slightly (it looked weird without window frames).
Fiddled a bit with the onscreen keyboard, so now the appropriate button labels change when a mod (Shift, Alt) button is pressed, nicer visuals.